refactor: split interfaces monolith into domain modules
Break the 1.2k-line interfaces/index.ts into domain files (user, site,
question, answer, article, etc.) with a barrel re-export so existing
`@/interfaces` imports keep working. Prefer new code to import from
specific modules when practical.
