+++ src/interfaces/audit.ts
@@ -1,0 +1,9 @@
+import type { IUserPreview } from './user';
+
+export interface IAuditLog {
+  uuid: string;
+  created_at: string;
+  api: string;
+  user: IUserPreview;
+}
+
